+++ effectful/src/Effectful/Prim/IORef.hs
@@ -66,6 +66,9 @@ atomicWriteIORef :: Prim :> es => IORef a -> a -> Eff es ()
 atomicWriteIORef var = unsafeEff_ . Ref.atomicWriteIORef var
 
 -- | Lifted 'Ref.mkWeakIORef'.
+--
+-- /Note:/ the finalizer will run a cloned environment, so any changes it makes
+-- to thread local data will not be visible outside of it.
 mkWeakIORef :: Prim :> es => IORef a -> Eff es () -> Eff es (Weak (IORef a))
 mkWeakIORef var f = unsafeEff $ \es -> do
   -- The finalizer can run at any point and in any thread.